add 11A51-ChineseHypothesis, 11P32-LevysConjecture
Thank you for your comment Thomas. 

The previous version of ChineseHypothesis missed the $ symbol to end the math mode. I havent noticed that since I was using just a plain text editor and not a colored Tex one. I also thought that the 2 step-compilation to cnl already checked the tex syntax. 

The files have been modified as you suggested: 
- correct the tex error. 
- use \dfn instead of deflabel 
- <80 characters per lines
- clean up some spaces.